moved data model into properly named namespace folder. Moved all classes into that namespace
