--- src/CodeAFriendDataModel/IronPythonInterpreter.cs
@@ -1,11 +1,0 @@
-using System;
-using System.Text;
-using System.Collections;
-using System.Collections.Generic;
-
-/// <summary>
-/// 
-/// </summary>
-public class IronPythonInterpreter
-{
-}
