--- src/CodeAFriendDataModel/LanguageStrategy.cs
@@ -1,24 +1,0 @@
-using System;
-using System.Text;
-using System.Collections;
-using System.Collections.Generic;
-
-/// <summary>
-/// 
-/// </summary>
-public abstract class LanguageStrategy
-{
-	/// <summary>
-	/// 
-	/// </summary>
-	public string Name { get; }
-
-	/// <summary>
-	/// 
-	/// </summary>
-	/// <param name="scriptBody"></param>
-	/// <param name="parameters"></param>
-	/// <returns>ScriptEvaluation</returns>
-	abstract public ScriptEvaluation Execute(string scriptBody, RuntimeParameters parameters);
-
-}
