--- src/CodeAFriendDataModel/RuntimeParameters.cs
@@ -1,26 +1,0 @@
-using System;
-using System.Text;
-using System.Collections;
-using System.Collections.Generic;
-
-/// <summary>
-/// 
-/// </summary>
-public class RuntimeParameters
-{
-	/// <summary>
-	/// 
-	/// </summary>
-	public uint MaxCpuTime { get; }
-
-	/// <summary>
-	/// 
-	/// </summary>
-	public uint MaxMemory { get; }
-
-	/// <summary>
-	/// 
-	/// </summary>
-	public string Input { get; }
-
-}
