--- src/CodeAFriendDataModel/Solution.cs
@@ -1,21 +1,0 @@
-using System;
-using System.Text;
-using System.Collections;
-using System.Collections.Generic;
-
-/// <summary>
-/// 
-/// </summary>
-public class Solution : Script
-{
-	/// <summary>
-	/// 
-	/// </summary>
-	public IEnumerable<Vote> Votes { get; }
-
-	/// <summary>
-	/// 
-	/// </summary>
-	public User Submitter { get; }
-
-}
