--- src/CodeAFriendDataModel/TestCase.cs
@@ -1,21 +1,0 @@
-using System;
-using System.Text;
-using System.Collections;
-using System.Collections.Generic;
-
-/// <summary>
-/// 
-/// </summary>
-public class TestCase
-{
-	/// <summary>
-	/// 
-	/// </summary>
-	public string Input { get; }
-
-	/// <summary>
-	/// 
-	/// </summary>
-	public string ExpectedOutput { get; }
-
-}
