--- src/CodeAFriendDataModel/User.cs
@@ -1,22 +1,0 @@
-using System;
-using System.Text;
-using System.Collections;
-using System.Collections.Generic;
-
-/// <summary>
-/// 
-/// </summary>
-public class User
-{
-
-	/// <summary>
-	/// 
-	/// </summary>
-	public IEnumerable<UserScript> Scripts { get; }
-
-	/// <summary>
-	/// 
-	/// </summary>
-	public string username { get; }
-
-}
