--- src/CodeAFriendDataModel/UserScript.cs
@@ -1,22 +1,0 @@
-using System;
-using System.Text;
-using System.Collections;
-using System.Collections.Generic;
-
-/// <summary>
-/// 
-/// </summary>
-public class UserScript : Script
-{
-
-	/// <summary>
-	/// 
-	/// </summary>
-	public string ScriptName { get; }
-
-	/// <summary>
-	/// 
-	/// </summary>
-	public string UserName { get; }
-
-}
