--- src/CodeAFriendDataModel/Vote.cs
@@ -1,27 +1,0 @@
-using System;
-using System.Text;
-using System.Collections;
-using System.Collections.Generic;
-
-/// <summary>
-/// 
-/// </summary>
-public class Vote
-{
-
-	/// <summary>
-	/// 
-	/// </summary>
-	public short Value { get; }
-
-	/// <summary>
-	/// 
-	/// </summary>
-	public User User { get; }
-
-	/// <summary>
-	/// 
-	/// </summary>
-	public string Comment { get; }
-
-}
