+++ src/shared/CodeAFriend.DataModel/CodaAFriendController.cs
@@ -1,0 +1,20 @@
+using System;
+using System.Text;
+using System.Collections;
+using System.Collections.Generic;
+
+namespace CodeAFriend.DataModel
+{
+	/// <summary>
+	/// 
+	/// </summary>
+	public abstract class CodaAFriendController
+	{
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public ICodeAFriendRepository Repository { get; }
+
+	}
+}
