+++ src/shared/CodeAFriend.DataModel/RuntimeParameters.cs
@@ -1,0 +1,29 @@
+using System;
+using System.Text;
+using System.Collections;
+using System.Collections.Generic;
+
+namespace CodeAFriend.DataModel
+{
+	/// <summary>
+	/// 
+	/// </summary>
+	public class RuntimeParameters
+	{
+		/// <summary>
+		/// 
+		/// </summary>
+		public uint MaxCpuTime { get; }
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public uint MaxMemory { get; }
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public string Input { get; }
+
+	}
+}
