+++ src/shared/CodeAFriend.DataModel/ScriptController.cs
@@ -1,0 +1,55 @@
+using System;
+using System.Text;
+using System.Collections;
+using System.Collections.Generic;
+
+namespace CodeAFriend.DataModel
+{
+	/// <summary>
+	/// 
+	/// </summary>
+	public class ScriptController : CodaAFriendController
+	{
+
+		/// <summary>
+		/// 
+		/// </summary>
+		/// <param name="script"></param>
+		/// <returns>UserScript</returns>
+		public UserScript CreateScript(UserScript script)
+		{
+			throw new Exception("The method or operation is not implemented.");
+		}
+
+		/// <summary>
+		/// 
+		/// </summary>
+		/// <param name="script"></param>
+		/// <returns>UserScript</returns>
+		public UserScript UpdateScript(UserScript script)
+		{
+			throw new Exception("The method or operation is not implemented.");
+		}
+
+		/// <summary>
+		/// 
+		/// </summary>
+		/// <param name="id"></param>
+		/// <returns></returns>
+		public void DeleteScript(Guid id)
+		{
+			throw new Exception("The method or operation is not implemented.");
+		}
+
+		/// <summary>
+		/// 
+		/// </summary>
+		/// <param name="id"></param>
+		/// <returns>Script</returns>
+		public Script GetScript(Guid id)
+		{
+			throw new Exception("The method or operation is not implemented.");
+		}
+
+	}
+}
