+++ src/shared/CodeAFriend.DataModel/ScriptEvaluation.cs
@@ -1,0 +1,35 @@
+using System;
+using System.Text;
+using System.Collections;
+using System.Collections.Generic;
+
+namespace CodeAFriend.DataModel
+{
+	/// <summary>
+	/// 
+	/// </summary>
+	public class ScriptEvaluation
+	{
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public string Output { get; }
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public uint CpuTime { get; }
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public uint MemoryUsage { get; }
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public RuntimeParameters Parameters { get; }
+
+	}
+}
