+++ src/shared/CodeAFriend.DataModel/Solution.cs
@@ -1,0 +1,24 @@
+using System;
+using System.Text;
+using System.Collections;
+using System.Collections.Generic;
+
+namespace CodeAFriend.DataModel
+{
+	/// <summary>
+	/// 
+	/// </summary>
+	public class Solution : Script
+	{
+		/// <summary>
+		/// 
+		/// </summary>
+		public IEnumerable<Vote> Votes { get; }
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public User Submitter { get; }
+
+	}
+}
