+++ src/shared/CodeAFriend.DataModel/TestCase.cs
@@ -1,0 +1,24 @@
+using System;
+using System.Text;
+using System.Collections;
+using System.Collections.Generic;
+
+namespace CodeAFriend.DataModel
+{
+	/// <summary>
+	/// 
+	/// </summary>
+	public class TestCase
+	{
+		/// <summary>
+		/// 
+		/// </summary>
+		public string Input { get; }
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public string ExpectedOutput { get; }
+
+	}
+}
