+++ src/shared/CodeAFriend.DataModel/User.cs
@@ -1,0 +1,25 @@
+using System;
+using System.Text;
+using System.Collections;
+using System.Collections.Generic;
+
+namespace CodeAFriend.DataModel
+{
+	/// <summary>
+	/// 
+	/// </summary>
+	public class User
+	{
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public IEnumerable<UserScript> Scripts { get; }
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public string username { get; }
+
+	}
+}
