+++ src/shared/CodeAFriend.DataModel/UserController.cs
@@ -1,0 +1,35 @@
+using System;
+using System.Text;
+using System.Collections;
+using System.Collections.Generic;
+
+namespace CodeAFriend.DataModel
+{
+	/// <summary>
+	/// 
+	/// </summary>
+	public class UserController : CodaAFriendController
+	{
+
+		/// <summary>
+		/// 
+		/// </summary>
+		/// <param name="user"></param>
+		/// <returns>User</returns>
+		public User CreateUser(User user)
+		{
+			throw new Exception("The method or operation is not implemented.");
+		}
+
+		/// <summary>
+		/// 
+		/// </summary>
+		/// <param name="userName"></param>
+		/// <returns>User</returns>
+		public User GetUser(string userName)
+		{
+			throw new Exception("The method or operation is not implemented.");
+		}
+
+	}
+}
