+++ src/shared/CodeAFriend.DataModel/UserScript.cs
@@ -1,0 +1,25 @@
+using System;
+using System.Text;
+using System.Collections;
+using System.Collections.Generic;
+
+namespace CodeAFriend.DataModel
+{
+	/// <summary>
+	/// 
+	/// </summary>
+	public class UserScript : Script
+	{
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public string ScriptName { get; }
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public string UserName { get; }
+
+	}
+}
