+++ src/shared/CodeAFriend.DataModel/Vote.cs
@@ -1,0 +1,29 @@
+using System;
+using System.Text;
+using System.Collections;
+using System.Collections.Generic;
+
+namespace CodeAFriend.DataModel
+{
+	/// <summary>
+	/// 
+	/// </summary>
+	public class Vote
+	{
+		/// <summary>
+		/// 
+		/// </summary>
+		public short Value { get; }
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public User User { get; }
+
+		/// <summary>
+		/// 
+		/// </summary>
+		public string Comment { get; }
+
+	}
+}
